Split single_table.gpt model to base class
Split part of the single table gpt model as a base class
